pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-ollama>=1.0.0",
    "langgraph>=1.0.3",
    "neo4j>=6.1.0",
    "pydantic>=2.12.4",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "requests>=2.32.0",
    "typing-extensions>=4.15.0",
]
